👌 IMPROVE: Cleanup
Laying down foundation for future updates by separating the routes into meaningful files.
Renamed the CodeHighlight js file for clarity
Removed unnecessary console.log()'s and white spaces.
